refactor: merge group-member and group module
