--- src/group-members/group-members.service.ts
@@ -1,12 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
-import { GroupMember } from './group-members.entity';
-
-@Injectable()
-export class GroupMembersService {
-  constructor(
-    @InjectRepository(GroupMember)
-    private readonly groupMemberRepository: Repository<GroupMember>,
-  ) {}
-}
