+++ src/groups/entities/group.entity.ts
