+++ src/permissions/permissions.entity.ts
@@ -1,5 +1,5 @@
 import { Base } from 'src/common/base.entity';
-import { Group } from 'src/groups/groups.entity';
+import { Group } from 'src/groups/entities/group.entity';
 import { Namespace } from 'src/namespaces/namespaces.entity';
 import { Resource } from 'src/resources/resources.entity';
 import { User } from 'src/user/user.entity';
